restore older test files
